fix: :bug: temp fix for filter search
